+++ KVStoreAlgorithm.h
@@ -88,15 +88,15 @@ private:
 
     ReplicaType ConvertToReplicaType(string replicaTypeString);
 
-    size_t myPositionInTheRing();
-    size_t findfirstSuccessorIndex(size_t myPos);
-    size_t findSecondSuccessorIndex(size_t myPos);
-    size_t findfirstPredeccesorIndex(size_t myPos);
-    size_t findSecondPredeccesorIndex(size_t myPos);
-    void setHasMyReplicas(size_t succ_1, size_t succ_2);
-    void setHaveReplicasOf(size_t pred_1, size_t pred_2);
-    void sendMessageToUpdateReplicaInfoFromPrimary(const string &key, const string &keyValue, size_t successorFirstIndex, size_t successorSecondIndex);
-    void sendMessageToUpdateReplicaInfoFromSecondary(const string &key, const string &keyValue, size_t predeccesorFirstIndex, size_t successorFirstIndex);
+    int myPositionInTheRing();
+    int findfirstSuccessorIndex(int myPos);
+    int findSecondSuccessorIndex(int myPos);
+    int findfirstPredeccesorIndex(int myPos);
+    int findSecondPredeccesorIndex(int myPos);
+    void setHasMyReplicas(int succ_1, int succ_2);
+    void setHaveReplicasOf(int pred_1, int pred_2);
+    void sendMessageToUpdateReplicaInfoFromPrimary(const string &key, const string &keyValue, int successorFirstIndex, int successorSecondIndex);
+    void sendMessageToUpdateReplicaInfoFromSecondary(const string &key, const string &keyValue, int predeccesorFirstIndex, int successorFirstIndex);
     
     // stabilization protocol - handle multiple failures
     void stabilizationProtocol();
